Basic functionality complete
It can properly identify free and bound variables, as well as evaluate
expressions. There are still some cases where the error reporting is
rather poor, such as when the function in a function application isn't
really a function. There may also still be bugs regarding name clashes.

Lastly, the `-v` doesn't work yet. It is ignored.

It's an MVP though. The code is still a bit messy, but I'm about done
spending time on it for now.